Backport move of environment listener
